requirements.txt
python-telegram-bot>=20.0
gspread
oauth2client
PyYAML
pandas
matplotlib
APScheduler
yfinance
requests
python-dotenv
